fix: corrigir consumer RabbitMQ no email-service
- Configurar TypePrecedence.INFERRED no Jackson2JsonMessageConverter para
  resolver incompatibilidade de __TypeId__ entre backend e email-service
- Corrigir tipo dos campos numericos de Integer para Double nos DTOs
- Corrigir variavel de ambiente MAIL_FROM no docker-compose.yml

+++ email-service/src/main/java/com/example/email_service/dto/EmailEnvioMoedaEvent.java
@@ -4,7 +4,7 @@ public record EmailEnvioMoedaEvent(
     String professorEmail,
     String professorNome,
     String alunoNome,
-    Integer valor,
+    Double valor,
     String mensagem,
-    Integer saldoRestante
+    Double saldoRestante
 ) {}
