+++ email-service/src/main/java/com/example/email_service/dto/EmailResgateAlunoEvent.java
@@ -5,7 +5,7 @@ public record EmailResgateAlunoEvent(
     String alunoNome,
     String vantagemNome,
     String codigoCupom,
-    Integer custoMoedas,
-    Integer saldoRestante,
+    Double custoMoedas,
+    Double saldoRestante,
     String empresaNome
 ) {}
